# Alternative bi-lock implementation