--- a/src/sync/bilock.rs
+++ b/src/sync/bilock.rs
@@ -1,5 +1,4 @@
 use std::any::Any;
-use std::boxed::Box;
 use std::cell::UnsafeCell;
 use std::error::Error;
 use std::fmt;
@@ -7,7 +6,7 @@ use std::mem;
 use std::ops::{Deref, DerefMut};
 use std::sync::Arc;
 use std::sync::atomic::AtomicUsize;
-use std::sync::atomic::Ordering::SeqCst;
+use std::sync::atomic::Ordering::AcqRel;
 
 use {Async, Future, Poll};
 use task::{self, Task};
@@ -33,12 +32,14 @@ use task::{self, Task};
 #[derive(Debug)]
 pub struct BiLock<T> {
     inner: Arc<Inner<T>>,
+    state: usize
 }
 
 #[derive(Debug)]
 struct Inner<T> {
-    state: AtomicUsize,
-    inner: Option<UnsafeCell<T>>,
+    free_state: AtomicUsize,
+    tasks: [UnsafeCell<Option<Task>>; 2],
+    value: Option<UnsafeCell<T>>,
 }
 
 unsafe impl<T: Send> Send for Inner<T> {}
@@ -52,11 +53,12 @@ impl<T> BiLock<T> {
     /// tasks to be managed there.
     pub fn new(t: T) -> (BiLock<T>, BiLock<T>) {
         let inner = Arc::new(Inner {
-            state: AtomicUsize::new(0),
-            inner: Some(UnsafeCell::new(t)),
+            free_state: AtomicUsize::new(2),
+            tasks: [Default::default(), Default::default()],
+            value: Some(UnsafeCell::new(t)),
         });
 
-        (BiLock { inner: inner.clone() }, BiLock { inner: inner })
+        (BiLock { inner: inner.clone(), state: 0 }, BiLock { inner: inner, state: 1 })
     }
 
     /// Attempt to acquire this lock, returning `NotReady` if it can't be
@@ -68,54 +70,19 @@ impl<T> BiLock<T> {
     /// the locked value (and can be used to access the protected data). The
     /// lock is unlocked when the returned `BiLockGuard` is dropped.
     ///
-    /// If the lock is already held then this function will return
-    /// `Async::NotReady`. In this case the current task will also be scheduled
-    /// to receive a notification when the lock would otherwise become
-    /// available.
-    ///
     /// # Panics
     ///
     /// This function will panic if called outside the context of a future's
-    /// task.
-    pub fn poll_lock(&self) -> Async<BiLockGuard<T>> {
-        loop {
-            match self.inner.state.swap(1, SeqCst) {
-                // Woohoo, we grabbed the lock!
-                0 => return Async::Ready(BiLockGuard { inner: self }),
+    /// task, or if the lock is already held.
+    pub fn poll_lock(&mut self) -> Async<BiLockGuard<T>> {
+        unsafe {
+            self.state = self.inner.poll_lock(self.state);
+        }
 
-                // Oops, someone else has locked the lock
-                1 => {}
-
-                // A task was previously blocked on this lock, likely our task,
-                // so we need to update that task.
-                n => unsafe {
-                    drop(Box::from_raw(n as *mut Task));
-                }
-            }
-
-            let me = Box::new(task::current());
-            let me = Box::into_raw(me) as usize;
-
-            match self.inner.state.compare_exchange(1, me, SeqCst, SeqCst) {
-                // The lock is still locked, but we've now parked ourselves, so
-                // just report that we're scheduled to receive a notification.
-                Ok(_) => return Async::NotReady,
-
-                // Oops, looks like the lock was unlocked after our swap above
-                // and before the compare_exchange. Deallocate what we just
-                // allocated and go through the loop again.
-                Err(0) => unsafe {
-                    drop(Box::from_raw(me as *mut Task));
-                },
-
-                // The top of this loop set the previous state to 1, so if we
-                // failed the CAS above then it's because the previous value was
-                // *not* zero or one. This indicates that a task was blocked,
-                // but we're trying to acquire the lock and there's only one
-                // other reference of the lock, so it should be impossible for
-                // that task to ever block itself.
-                Err(n) => panic!("invalid state: {}", n),
-            }
+        if self.state == 2 {
+            Async::Ready(BiLockGuard { inner: self })
+        } else {
+            Async::NotReady
         }
     }
 
@@ -149,33 +116,48 @@ impl<T> BiLock<T> {
         }
     }
 
-    fn unlock(&self) {
-        match self.inner.state.swap(0, SeqCst) {
-            // we've locked the lock, shouldn't be possible for us to see an
-            // unlocked lock.
-            0 => panic!("invalid unlocked state"),
-
-            // Ok, no one else tried to get the lock, we're done.
-            1 => {}
-
-            // Another task has parked themselves on this lock, let's wake them
-            // up as its now their turn.
-            n => unsafe {
-                Box::from_raw(n as *mut Task).notify();
-            }
+    fn unlock(&mut self) {
+        unsafe {
+            self.state = self.inner.unlock(self.state);
         }
     }
 }
 
 impl<T> Inner<T> {
     unsafe fn into_inner(mut self) -> T {
-        mem::replace(&mut self.inner, None).unwrap().into_inner()
+        self.value.take().unwrap().into_inner()
+    }
+
+    unsafe fn poll_lock(&self, caller_state: usize) -> usize {
+        assert!(caller_state != 2, "Lock already held");
+
+        // Save current task
+        *self.tasks[caller_state].get() = Some(task::current());
+        self.free_state.swap(caller_state, AcqRel)
+    }
+
+    unsafe fn unlock(&self, mut caller_state: usize) -> usize {
+        assert!(caller_state == 2, "Lock not held");
+
+        // Swap our state with the free state
+        caller_state = self.free_state.swap(caller_state, AcqRel);
+
+        // Wake up the other task
+        if let Some(task) = (*self.tasks[caller_state].get()).take() {
+            task.notify();
+        }
+
+        caller_state
+    }
+
+    unsafe fn get_value(&self) -> &mut T {
+        &mut *self.value.as_ref().unwrap().get()
     }
 }
 
 impl<T> Drop for Inner<T> {
     fn drop(&mut self) {
-        assert_eq!(self.state.load(SeqCst), 0);
+        assert_eq!(*self.free_state.get_mut(), 2);
     }
 }
 
@@ -210,19 +192,19 @@ impl<T: Any> Error for ReuniteError<T> {
 /// unlocked.
 #[derive(Debug)]
 pub struct BiLockGuard<'a, T: 'a> {
-    inner: &'a BiLock<T>,
+    inner: &'a mut BiLock<T>,
 }
 
 impl<'a, T> Deref for BiLockGuard<'a, T> {
     type Target = T;
     fn deref(&self) -> &T {
-        unsafe { &*self.inner.inner.inner.as_ref().unwrap().get() }
+        unsafe { self.inner.inner.get_value() }
     }
 }
 
 impl<'a, T> DerefMut for BiLockGuard<'a, T> {
     fn deref_mut(&mut self) -> &mut T {
-        unsafe { &mut *self.inner.inner.inner.as_ref().unwrap().get() }
+        unsafe { self.inner.inner.get_value() }
     }
 }
 
@@ -244,13 +226,16 @@ impl<T> Future for BiLockAcquire<T> {
     type Error = ();
 
     fn poll(&mut self) -> Poll<BiLockAcquired<T>, ()> {
-        match self.inner.as_ref().expect("cannot poll after Ready").poll_lock() {
+        match self.inner.as_mut().expect("cannot poll after Ready").poll_lock() {
             Async::Ready(r) => {
                 mem::forget(r);
             }
             Async::NotReady => return Ok(Async::NotReady),
         }
-        Ok(Async::Ready(BiLockAcquired { inner: self.inner.take() }))
+        let bi_lock = self.inner.take().expect("cannot poll after Ready");
+        Ok(Async::Ready(BiLockAcquired {
+            inner: Some(bi_lock.inner.clone())
+        }))
     }
 }
 
@@ -262,16 +247,17 @@ impl<T> Future for BiLockAcquire<T> {
 /// `unlock` method.
 #[derive(Debug)]
 pub struct BiLockAcquired<T> {
-    inner: Option<BiLock<T>>,
+    inner: Option<Arc<Inner<T>>>,
 }
 
 impl<T> BiLockAcquired<T> {
     /// Recovers the original `BiLock<T>`, unlocking this lock.
     pub fn unlock(mut self) -> BiLock<T> {
-        let bi_lock = self.inner.take().unwrap();
-
+        let mut bi_lock = BiLock {
+            inner: self.inner.take().unwrap(),
+            state: 2
+        };
         bi_lock.unlock();
-
         bi_lock
     }
 }
@@ -279,20 +265,20 @@ impl<T> BiLockAcquired<T> {
 impl<T> Deref for BiLockAcquired<T> {
     type Target = T;
     fn deref(&self) -> &T {
-        unsafe { &*self.inner.as_ref().unwrap().inner.inner.as_ref().unwrap().get() }
+        unsafe { self.inner.as_ref().unwrap().get_value() }
     }
 }
 
 impl<T> DerefMut for BiLockAcquired<T> {
     fn deref_mut(&mut self) -> &mut T {
-        unsafe { &mut *self.inner.as_mut().unwrap().inner.inner.as_ref().unwrap().get() }
+        unsafe { self.inner.as_ref().unwrap().get_value() }
     }
 }
 
 impl<T> Drop for BiLockAcquired<T> {
     fn drop(&mut self) {
-        if let Some(ref bi_lock) = self.inner {
-            bi_lock.unlock();
+        if let Some(inner) = self.inner.take() {
+            unsafe { inner.unlock(2); }
         }
     }
 }