--- a/tests/bilock.rs
+++ b/tests/bilock.rs
@@ -14,7 +14,7 @@ use support::*;
 #[test]
 fn smoke() {
     let future = future::lazy(|| {
-        let (a, b) = BiLock::new(1);
+        let (mut a, mut b) = BiLock::new(1);
 
         {
             let mut lock = match a.poll_lock() {
@@ -25,7 +25,6 @@ fn smoke() {
             *lock = 2;
 
             assert!(b.poll_lock().is_not_ready());
-            assert!(a.poll_lock().is_not_ready());
         }
 
         assert!(b.poll_lock().is_ready());
@@ -52,34 +51,38 @@ fn smoke() {
 
 #[test]
 fn concurrent() {
-    const N: usize = 10000;
-    let (a, b) = BiLock::new(0);
+    let _ = future::lazy(|| {
+        const N: usize = 10000;
+        let (a, b) = BiLock::new(0);
 
-    let a = Increment {
-        a: Some(a),
-        remaining: N,
-    };
-    let b = stream::iter_ok::<_, ()>((0..N)).fold(b, |b, _n| {
-        b.lock().map(|mut b| {
-            *b += 1;
-            b.unlock()
-        })
-    });
+        let a = Increment {
+            a: Some(a),
+            remaining: N,
+        };
+        let b = stream::iter_ok::<_, ()>((0..N)).fold(b, |b, _n| {
+            b.lock().map(|mut b| {
+                *b += 1;
+                b.unlock()
+            })
+        });
 
-    let t1 = thread::spawn(move || a.wait());
-    let b = b.wait().expect("b error");
-    let a = t1.join().unwrap().expect("a error");
+        let t1 = thread::spawn(move || a.wait());
+        let mut b = b.wait().expect("b error");
+        let mut a = t1.join().unwrap().expect("a error");
 
-    match a.poll_lock() {
-        Async::Ready(l) => assert_eq!(*l, 2 * N),
-        Async::NotReady => panic!("poll not ready"),
-    }
-    match b.poll_lock() {
-        Async::Ready(l) => assert_eq!(*l, 2 * N),
-        Async::NotReady => panic!("poll not ready"),
-    }
+        match a.poll_lock() {
+            Async::Ready(l) => assert_eq!(*l, 2 * N),
+            Async::NotReady => panic!("poll not ready"),
+        }
+        match b.poll_lock() {
+            Async::Ready(l) => assert_eq!(*l, 2 * N),
+            Async::NotReady => panic!("poll not ready"),
+        }
 
-    assert_eq!(a.reunite(b).expect("bilock/concurrent: reunite error"), 2 * N);
+        assert_eq!(a.reunite(b).expect("bilock/concurrent: reunite error"), 2 * N);
+
+        Ok::<(), ()>(())
+    }).wait();
 
     struct Increment {
         remaining: usize,
@@ -96,7 +99,7 @@ fn concurrent() {
                     return Ok(self.a.take().unwrap().into())
                 }
 
-                let a = self.a.as_ref().unwrap();
+                let a = self.a.as_mut().unwrap();
                 let mut a = match a.poll_lock() {
                     Async::Ready(l) => l,
                     Async::NotReady => return Ok(Async::NotReady),