This PR replaces the implementation of `BiLock` with one which does not perform allocation, ~and is slightly simpler~.

This introduces a backwards-incompatible change to the API as it requires a mutable reference when locking/unlocking a `BiLock`.

In general, lock-free data structures require space proportional to the number of concurrent accessors, so taking mutable references to `self` should be preferred as it allows more efficient implementations.

Implementation details:
- The `BiLock` is initialised with three "tokens", numbered 0, 1, 2.
- At all times, each side of the `BiLock` owns one token, and the remaining token is "free".
- Only one thread-safe operation is supported by a `BiLock`: atomically swapping its own token with the "free" token.
- Owning a token gives you exclusive access to part of the data structure: tokens 0, 1 gives access to the corresponding indices in the "tasks" array, while token 2 gives access to the value protected by the lock.

`poll_lock` operation:
- The current token must be 0 of 1, or else we'd already have the lock
- The current task can be stored in the `tasks` array at the element we have access to
- Atomically swap our token with the free token
- If our token is now "2", we own the lock
- Otherwise, the free token now corresponds to our own task

`unlock` operation:
- The current token must be 2
- Atomically swap our token with the free token
- We get back either the 0 or 1 token, which gives us access to an element in the tasks array
- Wake up the task we obtained from the token

Tests:
I had to wrap the "concurrent" tests in a lazy future, because they called `poll_lock` outside a task, which used to work despite the documentation saying that the method should panic when called off-task.