@Diggsey 2017-10-08
If this is of interest, a similar approach can be used for many of the other kinds of lock free data structure.