@Diggsey 2017-10-08
Benchmarks show marginally better performance for the contended case, but with significantly lower variation, which is about what I was hoping for by eliminating the allocation.

The contention-free case is showing exactly a 2x slow-down though, which is not good! I'm still investigating the cause, but initial evidence seems to point to a performance cliff (possibly a missed optimisation?) rather than an incremental slowdown caused by false sharing/increased size of `Inner`, so I'm hoping it's possible to fix!

```
Before:
test contended   ... bench:     177,635 ns/iter (+/- 52,813)
test lock_unlock ... bench:      59,665 ns/iter (+/- 4,362)

After:
test contended   ... bench:     177,250 ns/iter (+/- 13,123)
test lock_unlock ... bench:     117,321 ns/iter (+/- 8,443)
```