@Diggsey 2017-10-09
I discovered the major cause of slowdown was caused by storing the (mutable) token alongside the handle, rather than inside the data structure. Adding an extra layer of indirection improved the performance quite a bit:
```
Before:
test contended   ... bench:     177,635 ns/iter (+/- 52,813)
test lock_unlock ... bench:      59,665 ns/iter (+/- 4,362)

After:
test contended   ... bench:     144,758 ns/iter (+/- 12,908)
test lock_unlock ... bench:      77,714 ns/iter (+/- 6,243)
```

I may add some multi-threaded benchmarks, because at the moment all access is single-threaded and perfectly regular.