@omni-viral 2017-10-10
I believe state may be defined by single `AtomicUsize`
There are 6 states possible.
Here `left` and `right` are two owners.
```
s - [left,right,free]
0 - [0,1,2]
1 - [1,0,2]
2 - [2,0,1]
3 - [0,2,1]
4 - [1,2,0]
5 - [2,1,0]
```
Swap operation for `left` owner is `state.fetch_xor(1)`
for `right` is `state.fetch_add(3)`
to get current token id `STATE_MAP[state.load() % 6]` (and then `[0]` for `left`, `[1]` for `right`)