@Diggsey 2017-10-12
@omni-viral that's a really cool idea - I implemented it, but unfortunately it did not affect the performance. It should reduce the memory usage though.