@omni-viral 2017-10-12
@Diggsey replacing one read-modify-write operation with another shouldn't affect performance.
Atomic swap may be even faster than atomic add or atomic or on some platforms.