@Diggsey 2017-10-12
@omni-viral In the single-threaded test the bottle-neck was actually other factors than that atomic operations, possibly including the extra space, so I though it might help there.

I've added a concurrent test, and the changes do seem to be a significant improvement:
```
Before:
test concurrent  ... bench:   1,047,437 ns/iter (+/- 146,489)
test contended   ... bench:     161,603 ns/iter (+/- 11,326)
test lock_unlock ... bench:      54,580 ns/iter (+/- 4,425)

After:
test concurrent  ... bench:     695,118 ns/iter (+/- 118,187)
test contended   ... bench:     143,350 ns/iter (+/- 36,404)
test lock_unlock ... bench:      73,727 ns/iter (+/- 23,065)
```