@omni-viral 2017-10-13
@Diggsey I you don't mind can you run benchmark with state in single atomic and in separate and share results?