@alexcrichton 2017-10-19
Since this is a breaking change, would it be best to be in an external crate for now? One we could consider pulling in for the 0.2 release?