@Diggsey 2017-10-19
@alexcrichton I'm not sure `BiLock` has enough usage outside of the futures/tokio libraries themselves for there to be much benefit to publishing it in an external crate (in so far as it probably wouldn't get much extra testing prior to the 0.2 release). Was there another reason you suggested publishing in an external crate?

I'm in no particular rush to get this merged, but I do think using `&mut self` for these data structures is a change that should be made with 0.2 to avoid breakage down the line. (You can always wrap it in a `RefCell` to get back the previous semantics)