@arthurprs 2017-10-23
Interesting approach.

I wonder though if this is optimizing for the right use case? It appears to regress the single executor bench by ~38%, which is probably the most common due to split tcp sockets.