@Diggsey 2017-10-23
I'm not sure the lock/unlock bench is particularly representative of the common use-case either - in my experiments the performance was very dependant on LLVM's ability to optimise out the allocation. Even slight changes to the code which made no *semantic* difference could cause the performance to drop by as much as 2x, whereas the allocation-free version was much more consistent.