@carllerche 2017-10-23
Out of curiosity, has anyone actually compared `BiLock` with a plain ol' `Mutex`? It would be interesting to add that to the benchmarks.

Obviously a mutex blocks the thread, but if the critical section is small, that might end up just being the right solution.