@arthurprs 2017-10-24
@Diggsey I don't understand, the Box can't be optimized out. It must happen in order to return NotReady, but that's juts not the case for split sockets on the same executor.