@cramertj 2018-03-02
@Diggsey Would you be interested in rebasing this change and getting it in as part of the 0.2 release?