@Diggsey 2018-03-03
@cramertj I don't think this is ready to go in - I need to do some more bench-marking as there's definitely something unexplained with the performance characteristics.

However, the question of whether `self` should be mutable or immutable is important to answer for 0.2 - especially for `BiLock` I think it should take a mutable reference - it's very surprising that one side of a `BiLock` can be polled concurrently by multiple tasks, essentially making it a full (but unfair) mutex.