@cramertj 2018-04-03
I'm going to close this for now, but please reopen when you've had a chance to fix it up! Bilock [isn't reexported anymore](https://github.com/rust-lang-nursery/futures-rs/blob/master/futures-channel/src/lib.rs#L23), so we can fix this in the future.